[DocsUI, Search]: Fetch path on non-UI thread.
Accessing document path requires multiple content provider client
queries.  These should not be run on the UI thread. This CL extract
the code that makes the path request and construction of the IntConsumer
to a separate class. In addition, recent clicks are disabled, as for now
 they do nothing useful. A test is added.

Bug: 447678204
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests

Change-Id: I71a5d7dd958fa0800c9b1babc19e74a223dc3daf