[DocsUI Home screen] Edit breadcrumb bar to start at shortcut.
On a sidebar click, edit the breadcrumb bar so that it starts on a
shortcut rather than the local storage. Also ensure that this behaviour
remains consistent when we right-click on the "Home screen" sidebar
item and choose to open in new window.

Bug: 455712512
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FilesActivityUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.AbstractActionHandlerTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Flag: com.android.documentsui.flags.use_material3
Change-Id: I6e26fef6cb39d86f7003cb9859339f8ae430bbe9