[DocsUI, Search]: Fix breadcrumb path for multiple selection.
When one cannot find a unique ID for the selected search result, the
correct action is to set an empty path on the breadcrumb model. This
somehow was lost in the recent changes. This CL restores setting of the
empty path.

Bug: 455716850
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I150c7a61f5a4250d563bb3e415195e6584e275b6