[DocsUI] Fix locale configuration changes
When the language or locale is changed on the device, we expect
the sidebar roots and shortcuts on DocumentsUI to follow this change.
This CL updates the titles of shortcuts and the recents root to be
localised when the locale or language is changed.

Bug: 437259064
Test: atest DocumentsUIGoogleTests:com.android.documentsui.roots.ProvidersCacheTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Flag: com.android.documentsui.flags.use_material3
Change-Id: I9065b1ff7e6073355e3f8747cfc7b6b1a7355276