DocsUI: Update logic to display the Summary column
Update the column header and the column in RecyclerView to only show for
local roots or recents and when the summary provider is available and
enabled.

Change the logic to hide to use GONE instead of INVISIBLE so it doesn't
take any space.

Change to destroy the loader after it's finished, because it seems that
previous instances were lingering around.

Bug: 436750999
Flag: com.android.documentsui.flags.use_file_summary
Test: atest DocumentsUIGoogleTests
Change-Id: I9e714eb55deb0e74d5f9f3e96f0c2dc0c92eb3bf