[DocsUI, SearchV2]: Show breadcrumbv2 in search and recents.
Removes the condition for docked search. Regardless if the search is
docked, or not, if we are searching or in recents breadcrumb v2 should
be shown if search v2 is enabled. Also, we need to close not cancel
search for dropdowns to be replaced with chips.

Bug: 447678204
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I4827c39cca88e58b08c31f801f38944981a30cd2