Prevent a malicious selector from launching an arbitrary activity.
PickActivity reuses the incoming intent to show more apps that can handle the request. If the original intent has a selector, it can be used to launch an arbitrary activity with the permissions of DocumentsUI.
This change fixes the vulnerability by clearing the selector on the copied intent.

Bug: 447135012
Flag: EXEMPT BUGFIX
Test: manual
Change-Id: I3e2eeaab8990a20fe639a165630ed1773e47fb3c