[DocsUI Home screen] Change snackbar text for blocking file ops.
The snackbar text currently says "File operation not supported" which
is what happens when the file operation is rejected. On the UX PRD, it
says that the snackbar text should say "Action not allowed". We also
create a new string property for this even though
toast_action_not_allowed exists with the same text content so that it
the new text is exclusively for when a file operation is not allowed to
be executed on a shortcut.

Bug:456588367
Test: atest DocumentsUIGoogleTests:com.android.documentsui.AbstractActionHandlerTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.BlockFileOperationsForShortcutsUiTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Flag: com.android.documentsui.flags.use_material3
Change-Id: I48823979a88166ddad9bf909f902ff42af69df89