[DocsUI, Search]: Fix testRecentsShowsZipFiles
bots.directory.waitForDocument does not fail if waiting fails to find
the desired document. As a consequence this test passes by chance. The
fix is to move to the Downloads root before waiting for the newly
created file to appear.

Bug: 455731254
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I62c64aa948ea1420842d3075828a6dc405307506