[DocsUI, Search]: Fix checking of cursor staleness
BaseFileLoader is checking if moveToPosition(-1) was successful, before
traversing the cursor. However AbstractCursor always returns false on
moveToPosition(-1). Thus we always were under the impression that the
cursor was stale. This removes the check. This fixes the race conditon
where upon restoring the application it was possible to close the
results just returned to the caller (see bug for details).

Bug: 457590189
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I1b520fdd3e4678ba5cd0a8d62a68620985e0fd62