[DocsUI, Search]: Show V2 breadcrumb only on selection
Restrict the V2 breadcrumb visibility in Search and Recents views.
Previously, it would be shown as soon as search or the recents view
became active.

Move the control logic to DirectoryFragment to ensure the breadcrumb
only displays when a specific file is selected, accurately showing
the path for that individual item.

Bug: 437789231
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I00ab7546de708b16abe031fd3fe669172077ad77