[DocsUI] Fix failing CopyLargeAmountOfFiles test.
Test was failing due to an assertion thrown when waitForDocument
is unable to find a copied file. This occurred because it is not
in the view without scrolling to the bottom of the directory.

Added scrolling into test and modified number of swipes before
the test would return early.

Bug: 455729551
Test: atest com.android.documentsui.FileManagementUiTest#testCopyLargeAmountOfFiles
Flag: EXEMPT BUGFIX
Change-Id: I0b2d02c816ca07304dac8f9915dc0b4576cfe23a