[DocsUI] proguard BaseActivity.getLayoutId
Currently, some ABTD test runs (admittedly based on prebuilts) are
failing with NoSuchMethodError re getLayoutId, a method that is declared
in the src/ code but called only in the tests/ code. This proguard
change may help that, and probably doesn't hurt, whenever the prebuilts
are next refreshed.

Bug: 437215252
Flag: EXEMPT BUGFIX
Test: atest DocumentsUIGoogleTests
Change-Id: I4431aa43d442a68fb73bcf5e5128f544c10514c4