Fix null pointer when pasting image on clipboard
When an image is put on the clipboard and pasted into a folder, if it
originated outside of DocumentsUI it doesn't contain an operation type.
This causes a null pointer when attempting to get the op type from the
Bundle. Let's ignore this (instead of crash) for now as it's currently
not supported.

Fix: 458273555
Test: atest com.android.documentsui.clipping.RuntimeDocumentClipperTest
Flag: EXEMPT BUGFIX

Change-Id: Idee91bfeda9c46a816b773064bc8d197330ac073