[DocsUI] Disable open for disabled documents from the context menu
Open should always be disabled on disabled documents, including on a
selection containing disabled documents, as defined in this document
https://docs.google.com/document/d/1kZ6ifpo4ee3YpYzX8dAMSxr2AiTQOqMfKhMQMfh48Bo/edit?tab=t.sywrkzbdkwr8

Opens from double clicking are already disabled however the "Open with" menu item was not
so disabled documents could still be opened. This is because DocumentsUI previously was just
used as a file picker, instead of also a file manager, and pickers didn't use context menus.

This CL disables but still shows a greyed out "Open with" menu item in the context
and action menus for selections that contain disabled files.

Flag: com.android.documentsui.flags.use_material3
Bug: 459263894
Test: atest com.android.documentsui.dirlist.SelectionMetadataTest
Test: atest com.android.documentsui.files.MenuManagerTest
Change-Id: Idbfdc165def8389c67111918939b298e82890fa5