Handle race condition when mDestRoot is null
Currently there is a race condition that happens when a drag is
initiated with the Ctrl key being held down. It ends up attempting a
null dereference which crashes DocsUI. This handles the case and adds
some additional annotation around nullability of the offending member
variables.

Fix: 442450613
Test: atest com.android.documentsui.DragAndDropManagerTest
Flag: EXEMPT TEST_ONLY
Change-Id: Ia433571172317d34ff512b0a058d869ef2e62a08