Reduce timeout for finding menu items to 3s
A previous CL (ag/36867837) introduced a method to scroll the context
menu if an item is not found. It used mTimeout which is 15s to do this.
This is probably a bit aggressive given it will most likely get
triggered in some tests on the happy path (i.e. things WILL be hidden).
Let's reduce the timeout to 3s to ensure when something is hidden it
doesn't take too long to initiate the scroll.

Bug: 450206203
Bug: 434905747
Bug: 450839013
Test: atest com.android.documentsui.ContextMenuUiTest
Flag: EXEMPT TEST_ONLY
Change-Id: I4aaac75ec18675cf9a30b6bb3fc8dc080346b889