Return false instead of a throw when scrolling for menu items
A previous CL added menu scrolling to test if a menu item is available
or not. Unfortunately even if a scroll happens, the menu item might
still not be available. In that case the getChildByText method throws an
error instead of returning false and having the test assert that is the
required state. Update the method to return false as expected.

Bug: 450206203
Bug: 434905747
Bug: 450839013
Bug: 459620788
Test: atest com.android.documentsui.ContextMenuUiTest
Flag: EXEMPT TEST_ONLY
Change-Id: I8b5c9844d1d2e7a3b2252dffa4c8e497d28f29ab