[DocsUI] add show_selection_checkmark
This is a new overlayable config boolean that, when combined with the
single_click_to_select flag, lets us hide the selection checkmarks on
desktop (mouse-oriented) configurations.

The boolean's default value (true) means no change in behavior, when
comparing before and after this commit.

Also clarify the DirectoryListBot.selectDocument semantics (a method in
the test code). This does not change the existing semantics.

Also update its implementation to use a "long finger click" instead of a
"click", since there is no longer guaranteed to be "a selection hotspot"
(if show_selection_checkmark is false then the list-item's thumbnail
still exists but is no longer a hotspot and the grid-item's selection
circle simply no longer exists). A bare mouse click (on what was the
hotspot) is no longer guaranteed to toggle selection (instead of to set
one item's selectedness and clear other items' selectedness).

Bug: 427858500
Flag: com.android.documentsui.flags.single_click_to_select
Test: atest DocumentsUIGoogleTests
Change-Id: I76b4b01f510b32b30d20ecfe38282b144fbf1b02