[DocsUI] Add scrolling into browseInvalidArchiveViaContextMenu test.
Previously this test was flaky and would fail on certain devices where
the form factor did not have the target string on the screen without
scrolling. Added in scrolling to the test to ensure it succeeds
regardless of form factor. This was accomplished by refactoring some
common scroll logic into a helper function.

Bug: 458110351
Test: com.android.documentsui.ArchiveUiTest#browseInvalidArchiveViaContextMenu
Flag: EXEMPT BUGFIX
Change-Id: I22ce527e7a18fa3b6de4f6bb95fda0b79c1aaa2f