[DocsUI] Use Espresso scrollTo and actionOnItem in actionOnRootItem.
PickActivityTest suite was experiencing flaky tests because the
RecyclerView containing the roots list was being updated and
redrawn in an non-atomic way. There was no synchronization
to prevent performing an action on an item that did not exist
during this update. An internal class ActionOnRecyclerViewItem
was being used to prevent the frequency of this issue, but it
was still occurring. It has been resolved by adding in a wait
until completely displayed check on the view. The Espresso
actionOnItem method can then be used to scroll and select the root.

Bug: 456630768
Test: com.android.documentsui.PickActivityTest#testPickFilesFragment_ActionOpenDocument_SingleFile
Flag: com.android.documentsui.flags.use_material3
Change-Id: Iea72d3bec3419eb63d87b4c9aa1e413ce6dc7a87