Stub async operations in `DropOperation` helper class.
This CL stubs async operations in the `DropOperation` helper class
for rewriting clip data and calculating op type. Note that the actual
async implementations will come in a follow up CL and be flag guarded
to only take effect for drags from other apps.

This CL is a pure refactor and does not change existing behavior.

Bug: 440196110
Flag: EXEMPT PURE_REFACTOR
Test: atest DragAndDropManagerTests
Change-Id: Ied722e37c7f4b9893978e32977357924c43ef036