Introduce a functional test for the Get info dialog plumbing
The GetInfoDialogFragment contains majority of the logic around data
retrieval for the dialog. However, there is still a bit of (albert
fairly static) logic for launching the dialog. To ensure test coverage
for this scaffolding, introduce a simple UI test that just verifies the
end to end flow of opening a real file and asserting the info is
present.

Bug: 458494432
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.GetInfoDialogUitest
Change-Id: I9a95285ce8c5f1e78133b3b6d8e121c44d828ee4