[DocsUI] Handle when cursor's int column has a null value
It is possible that an int column on a cursor can have a null
value. For example, the COLUMN_ICON in DocumentsContract is an
int that can be null.

Currently our getCursorInteger function relies on the Cursor
getInt() method. However this returns 0 whenever the value is
null [1]. Instead we should be returning our own default value
(returnIfMissingOrNull).

Use the Cursor isNull method to check for this case and return
our own default value.

Also fix the RootCursorWrapper.isNull method, which gets called
within the chain of isNull() calls that gets to the MatrixCursor.isNull,
and ensure that only indexes that exist are passed into the next isNull.

[1] https://android.googlesource.com/platform/frameworks/base/+/master/core/java/android/database/MatrixCursor.java#283

Flag: EXEMPT BUGFIX
Bug: 463217744
Test: atest com.android.documentsui.base.DocumentInfoTest
Test: atest com.android.documentsui.AbstractActionHandlerTest#testCrossProfileDocuments_reloadSuccessAfterCrossProfileError
Change-Id: I232fd8a732f2a1ed88abdce309ce90d4c511918c