[DocsUI M3] Move some actions to the AbstractActionHandler
Previously some actions (e.g. Delete/Rename/Inspect) are only supported
in the files/ActionHandler but not picker/ActionHandler. In order to
support these actions inside file picker, this CL moves some actions
to the base class AbstractActionHandler:
* The corresponding tests are also moved to base class's test.
* The arugments of the base class has updated to cater the newly added
  methods.

Some logic defined in the FilesActivity also get moved to BaseActivity
in order to be shared with PickActivity.

Note: in order to put this behind the flag, picker/ActionHandler only
calls the super method when the flag is ON.

Bug: 461645421
Test: atest DocumentsUIUnitTests:com.android.documentsui.AbstractActionHandlerTest
Test: atest DocumentsUIUnitTests:com.android.documentsui.files.ActionHandlerTest
Test: atest DocumentsUIUnitTests:com.android.documentsui.picker.ActionHandlerTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ife6b08d51935716e473edd3a4afa68023d021453