Fix: Allow restoring files from a trashed folder
This change fixes a bug where restoring a file would fail if its parent folder was also in the trash.
The issue occurred because the restore operation was attempting to use the trashed parent folder as the destination. The fix checks if the current location is the trash folder. If it is, it sets the destination to null, allowing the system to restore the file to its original location.

Bug: 463670278
Flag: EXEMPT BUGFIX
Test: atest DocumentsUIGoogleTests:com.android.documentsui.TrashViewUiTest#testRestoreFileFromTrashedFolder
Change-Id: I366f343a747a8d33fb8c5de832deacd244575d3a