[DocsUI tests] Use NetworkMonitorStub in TestEnv
Previously a real NetworkMonitorImpl was being used (when the
CloudFeatures feature flag was enabled). This is an issue because
there is no TestEnv teardown method, and so the NetworkMonitorImpl
was never torn down. As then unregisterNetworkCallback was never called
this led to `android.net.ConnectivityManager$TooManyRequestsException`
eventually when many tests were run. This can be reproduced with
`atest DocumentsUIGoogleTests:com.android.documentsui.files.ActionHandlerTest`.

As the real NetworkMonitorImpl is not required in TestEnv, use the NetworkMonitorStub
instead.

The alternative would be to introduce a teardown TestEnv method and call it
everywhere. But this is a huge change compared to this simple fix.

Flag: com.android.documentsui.flags.cloud_features
Bug: 451775371
Test: atest com.android.documentsui.files.ActionHandlerTest
Change-Id: Ice908ef21968513643a587849d41c5079ab69f90