[DocsUI] Handle MediaStore URI intents from the Launcher.
As part of the Files on Home screen project, folders on the Launcher
home screen can be opened and viewed. The expected behaviour is to
open the folder's contents on DocumentsUI. However, when opening from
the Launcher, the intent data URI is a MediaStore URI which is not
recognisable to DocumentsUI. Use MediaStore.getDocumentUri() to convert
this MediaStore URI into a Documents URI so that DocumentsUI can open
the correct folder rather than a default folder location.

Bug: 441860560
Bug: 449845678
Flag: com.android.documentsui.flags.home_screen_files_ro
Flag: com.android.documentsui.flags.use_material3
Test: atest DocumentsUIGoogleTests:com.android.documentsui.files.ActionHandlerTest
Change-Id: I418467a0915679d9f544ae708bfb4991b1a73f6c