[DocsUI] Restyle icon in banner
A banner consists of an icon, a message container and a button.
The message container has a minimum height, so it will be filled with
whitespace if the text inside is not long.

Previously the icon would be centred vertically on the message
container. It should be centred on the actual text within the
container. Rearrange the banner xml elements to achieve this.

Flag: com.android.documentsui.flags.use_material3
Bug: 405853843
Test: EXEMPT layout change
Change-Id: I2dc548918ce7136492850892636e8e58a2090571