Ignore testSearchV2LastUsedChipCopiedToFileTypeDrown
This test is constantly failing now, given time prioties let's disable
it as it's super close to the the next train boarding and the flag
(`USE_SEARCH_V2_READ_ONLY`) is targeted to a later release.

This will provide additional time to find a more robust fix.

Bug: 454313609
Test: atest com.android.documentsui.SearchViewUiTest
Flag: EXEMPT TEST_ONLY
Change-Id: Ib84cd79a95b2fdb12fa891660aed0df6830bc9fb