Use DisableFlags instead of RequireFlagsDisabled
When the flag has rolled out, it seems the DUT has the flag disabled,
but the test seems to enable the flag causing the inspector tests to run
and go down the GET_INFO_DIALOG flag flow. Explicitly disable them in
the inspector tests to ensure the flag is at the expected state.

Bug: 458494432
Bug: 465516352
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.files.ActionHandlerTest
Change-Id: Ib7d9dc4b928bfcc032f0f6227abb5638e6ca0160