[DocsUI] control-space toggles selection
This commit is a no-op unless the use_material3 flag is enabled.

Also stop control-enter, shift-enter, etc (as opposed to a bare enter)
from opening a file.

Also stop arrow-keys-to-move-focus from clearing selection. It's not
clear why the previous behavior was implemented that way, but we no
longer clear the selection (if use_material3) so that we can now combine
arrow-keys and control-space to make a non-contiguous selection.

Also auto-format some surrounding KeyInputHandler.Callbacks code.

Bug: 417602656
Bug: 417872166
Flag: com.android.documentsui.flags.use_material3
Test: atest DocumentsUIGoogleTests
Change-Id: I53b66b44bcec996df25ad7710bc4be941493180c