[DocsUI M3] Prevent soft keyboard overlap with UI
In saver mode the file name input is at the bottom, when the soft
keybobard pops up (e.g. on Tablet), it will overlap with the UI and
cover the input field, so the user couldn't see what they are typing.
This CL adds the IME insets into consideration to fix the issue.

Check the attached bug for the demo.

Bug: 465198552
Test: m DocumentsUIGoogle && manual test
Flag: com.android.documentsui.flags.use_material3
Change-Id: I3e2a5a7b570458fb76bb18b1f8fd79b0d694c870