Add the debug information to the bottom of the Get info dialog
There is some synchronous debug information that we should show in the
Get info dialog (only on debug builds). This adds that functionality and
updates the asynchronous data to ensure it stays in position instead of
getting bumped to the bottom of the list.

Bug: 458494432
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.files.get.info.GetInfoViewModelTest
Change-Id: Icd7c37ceab680a6fbba9dd2ac2d5957e87437420