Refactor the GetInfoViewModel to use declarative Flows
The current setup in the Get info dialog is well suited for static data.
Unfortunately there are a number of upcoming data sources (EXIF data,
stream types etc.) that do not neatly fit in this setup. The brittle
setup of adding a ListItem then enumerating and updating that ListItem
when the async data is availble is not sustainable when a single query
might return N results where N is not known up front.

To get around this, refactor the items to instead map (but later
combine) Flows into a single StateFlow. This allows for the position of
the elements in the list to be set when the flows are updated and retain
their position. This has updated the tests as well to move towards
using an UnconfinedTestDispatcher so that the tests are all ran eagerly.
A few test helpers have been added to make things easier to read.

Bug: 458494432
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.files.getinfo.GetInfoViewModelTest
Change-Id: I3675502b4f702dd11f5904304884a4eba7375243