Explicitly check for `forceUriPermissions` during drop handling.
This CL explicitly checks for the `forceUriPermissions` attribute during
drop handling. Previously we relied on an assumption that the attribute
was enabled for the MediaProvider.

Note that this allows us to handle drops from any DocumentProvider which
enables the `forceUriPermissions` attribute which is an improvement over
only accepting MediaStore URIs.

Bug: 440196110
Flag: com.android.documentsui.flags.drags_from_other_apps
Test: atest DragAndDropManagerTests
Change-Id: Ib84907f6b807639fced59609d2e3e48c73d8533a