[DocsUI] Show inline tick after upload or download finishes
Show tick on DocumentHolder item for 1.2s after an upload or download
finishes. This tick takes the least priority over the other inline sync
icons, so only show it if no other icons could be shown.

Implement this by tracking which items are in an upload or download and
observing when the sync has finished. At this point, the modelId of the
item should be stored in a mJustFinishedSyncingModelIds in
ModelBackedDocumentsAdapter.

Whenever the DocumentHolder gets bound, pass on whether the item is in
this set. The holder can then show the tick.

To ensure that the tick disappears after 1.2s, a task should be
scheduled to remove the item from the set and trigger a rebind
(through notifyItemChanged).

See bug for demo.

Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.cloud_features
Bug: 463161690
Test: atest com.android.documentsui.dirlist.DocumentHolderTest
Test: atest com.android.documentsui.dirlist.ModelBackedDocumentsAdapterTest
Test: atest com.android.documentsui.ModelTest
Change-Id: I64746b2655132c78156906b45e7843913eed310e