Don't show folder modification times in grid view
This ensures that the grid view is consistent with the list view on this
point. At the moment, the list view doesn't show any modification times
for folders either.

The modification time of a folder also needs to be better defined before
we can decide to display it. See the discussion in bug 424686673.

Bug: 424686673
Bug: 452682985
Flag: com.android.documentsui.flags.use_material3
Test: atest DocumentsUIGoogleTests:com.android.documentsui.dirlist
Change-Id: Id25b83f72af333fe4cabd57a3186291766e2fc26