[DocsUI] classifySelectionHotspot handles ctrl-click
This commit should have no change in behavior, but it prepares for a
follow-up commit (in a separate repository) where the responsibility for
"control-click means SELECTION_HOTSPOT_INSIDE_TOGGLE_MULTI" moves from
the classifySelectionHotspot caller (i.e. code in the androidx
recyclerview-selection library) to the callee (which, in our case, is
code in the DocumentsUI app), to give apps more options on how to handle
control-click.

Bug: 389814214
Bug: 429096800
Flag: com.android.documentsui.flags.use_material3
Test: atest DocumentsUIGoogleTests
Change-Id: I0d1d1cfff99eba5bde30fcc9cfa7fa14bccd8793