Ensure the onSearchChanged callback doesn't get called on constructor
A change was landed in ag/36248687 that changed how the underlying
SearchViewManager works. Unfortunately changing the mCurrentSearch to
use setCurrentSearch in the SearchViewManager constructor causes the
onSearchChanged to be called on the BaseActivity before the activity
has been properly initialized. This leaves the Search in a weird
middle state.

This CL reverts the change in the constructor to the old code which
confirms it works. This was partially fixed in ag/37366496 but missed
the use case where the search query is empty. Along with that it adds a
new test that fails without the constructor change and passes with it.
This should help avoiding the scenario again in the future.

Bug: 466965174
Test: atest com.android.documentsui.SearchViewUiTest
Flag: EXEMPT BUGFIX
Change-Id: I2cb47814b7988e5579917eef82858e7a250e7f76