Add the result from getStreamTypes to Get info debug dialog
Stream types are an asynchronous value that is retrieved from the
DocumentsProvider. Its value represents the mimeType that the file will
take if the item is copied from the source DocumentsProvider. For
example if you have a bespoke document type, you can say that when a
openFile is called on it, it may return an application/pdf and the DP
will take care of the conversion.

Bug: 458494432
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.files.getinfo.GetInfoViewModelTest
Change-Id: I83e74e6b5639939d04785677930eb3ca2cd34c8c