[DocsUI M3] Prevent unexpected focus on left-top root item
The root container's focusability was removed in CL:36018336 to avoid
the grey-ish overlay. However, after that the system will try to focus
the left-top view (the "Recent" root item in the sidenav) when key press
is ignored in some scenarios (e.g. press Enter/Space when nothing is
focused, press any A to Z characters), even if we want to manually
handle the focus, the focus goes to "Recent" root first and then goes to
the viw we are trying to focus, which causes a confusing UX.

This CL brings the focusability of the coordinator_layout back, adds a
custom background which overrides the default focused background to
avoid the grey-ish overlay. So the system will try to focus it instead
but without showing the annoying grey-ish overlay. Note: it also
requires the attribute "descendantFocusability=beforeDescendants" on
the container to make sure it gets the focus before all descendants
when nothing is focused, otherwise it will still try to focus "Recent",
this will impact the tab navigation a bit but it's better than randomly
focusing the "Recent" root. One test is updated becaue now it takes
more than 1 tab press to focus the root list.

Bug: 454187623
Test: m DocumentsUIGoogle && manual test
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ib23c99ce750b3cea8dd36b50bcf3493e24f6ae30