[DocsUI, Search]: Show Last Modified dropdown in Recent.
If the Last Modified dropdown is not shown in Recent, it is possible for
the user to lock themselves in other than last month last modified
option. To avoid this, we are now showing the last modified dropdown in
Recents, regardless of the location searched.

Bug: 469886434
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: Ic0fdd824e34a3e0a25301e74f7be0da3c57e0669