[DocsUI] pick a different ContextMenuUiTest test file
This is a test only change.

Prior to this commit, testContextMenu_onMixedFileDir would try to select
two files: "file1.png" and "Dir1".

However, ag/36903447 "add show_selection_checkmark" added a
configuration option to show or hide the selection checkmark, and
"clicking on the selection checkmark (to select something)" would
obviously not work if show_selection_checkmark was false. To avoid
trying to "click on the selection checkmark", the way the test code
selected files changed. From ag/36903447's commit message:

...

Also clarify the DirectoryListBot.selectDocument semantics (a method in
the test code). This does not change the existing semantics.

Also update its implementation to use a "long finger click" instead of a
"click", since there is no longer guaranteed to be "a selection hotspot"
(if show_selection_checkmark is false then the list-item's thumbnail
still exists but is no longer a hotspot and the grid-item's selection
circle simply no longer exists). A bare mouse click (on what was the
hotspot) is no longer guaranteed to toggle selection (instead of to set
one item's selectedness and clear other items' selectedness).

...

Long finger click does indeed trigger selection, except that on
relatively small screens, the position of the file1.png document in
DocumentUI's grid view can be coincidentally almost off-screen, at the
bottom of the screen, and a long press (instead of a click) can
accidentally trigger Circle To Search (instead of selecting the file).

This commit changes "select (by long finger click) file1.png" to "select
(by long finger click) anotherFile0.log", since "anotherFile0.log" is
less likely (when the grid view is sorted by name) to be exactly under
the Circle To Search trigger zone.

There are other tests within ContextMenuUiTest that manipulate file1.png
but they call bots.directory.rightClickDocument on it, not
bots.directory.selectDocument.

This commit is something of a temporary fix that simply makes the test
green. In the long term, if show_selection_checkmark becomes 'hard
coded' instead of a configuration value, then it might be worth
revisiting "tests use long finger click to select". Or additional code
to "scroll the document into view before trying to select it". Or
something else. But anything like that is more complicated than this
one-line quick fix.

Bug: 427858500
Bug: 465325630
Flag: EXEMPT TEST_ONLY
Test: atest DocumentsUIGoogleTests:com.android.documentsui.ContextMenuUiTest#testContextMenu_onMixedFileDir
Change-Id: I541a17d3542164de92b6d56a015fb6c549b269cc