[DocsUI, Search]: Fix root change handling.
The reaction to the root change did not update the state's stack until
half way through the root change. As a consequence, if any code that was
called retrieved stack's root from the state, it got the old value. This
change sets the new root immediately. To preserve existing functionality
it also keeps record of old root. It adds a test to verify the new,
corrected behavior.

Bug: 437804247
Flag: EXEMPT BUGFIX
Test: atest DocumentsUIGoogleTests
Change-Id: I6746f694c2be10b550bb60cab85068419b154e39