[DocsUI Shortcuts] Hide DownloadStorageProvider sidebar item
If a shortcut to ExternalStorageProvider's "Download" folder is
provided, then hide the sidebar item for the DownloadStorageProvider
root.

Using the ExternalStorageProvider will fix a lot of minor issues
with the DownloadStorageProvider root and will enforce consistent
behaviour across the local storage items on the sidebar.

This CL will be submitted along with ag/37551548.

Bug: 469583396
Flag: com.android.documentsui.flags.home_screen_files_ro
Test: m DocumentsUIGoogle && manual inspection
Change-Id: I9fa3911c5dc8e42161a2c8c1ac7507883f895ffb