Update SDK version check for isContentProviderClientAnrOnCancelEnabled
When the flag check is present, in theory the SDK check should not be
necessary, as the flag should be correctly evaluated to false on all
Android versions other than ToT. Unfortunately, sometimes the flag
appears to be incorrectly evaluated to true when it shouldn't, leading
to missing API method crashes on older Android versions where
DocumentsUI is installed as a mainline module. Some of these crashes
were caused by incorrect use of the flag library (see ag/37462314),
but even with the fix it still crashes on Android U. While the root
cause of the last crash is being investigated, just limit the flag
check to Android B for now.

Bug: 409431618
Bug: 457843307
Test: start a manual copy of 1GB+ uncached file out of Drive + cancel
Flag: android.content.flags.enable_content_provider_client_anr_on_cancel
Change-Id: Ie3dcbe99441bc6f57b2d21ce09894ab2adb79a9e