Fix up flakiness in GetInfoViewModelTest
There are some subtle race conditions in the GetInfoViewModelTest that
cause it to always fail when ran with --rerun-until-failure 100. The
existing Channel-based flow collection and dispatcher configuration can
cause the consumer to beat the producer and assert on a previous state
of the list.

To get around this, fall back to the StateFlow.first predicate which
builds in the automatic subscription of a flow and suspends the flow
until the predicate is satisfied. This coupled with having the ViewModel
background work run on an eager execution coroutine (i.e.
UnconfinedTestDispatcher) and having the test run on a
StandardTestDispatcher allows for the first to properly suspend and wait
for the flow to execute.

Along with this, introduce some helper methods that make the error
assertions much more descriptive. Avoid testing intermediate states as
well, upcoming changes will make the intermediate states harder to
reason about (e.g. 3 async flows running in parallel can't always
deterministically assert on the sequence of one finishing over the
other).

Bug: 458494432
Flag: com.android.documentsui.flags.get_info_dialog
Test: atest com.android.documentsui.files.getinfo.GetInfoViewModelTest
Change-Id: I08cfc727a8096a4f7c4373cc5e90b7e5a3ca7b52