[DocsUI] Limit refresh root and directory call from loaders.
Only call to refresh the current root and directory via loaders
when the reloadRootsAndShortcuts call is directly coming from
BaseActivity.onConfigurationChanged().

This will be done by passing in a boolean and then bundling this up
as args and then refreshing the current root and directory based on
the boolean value from the bundle.

Bug: 469318781
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.home_screen_files_ro
Change-Id: I4ede302b139a0aeaaf1b2064d97f208971f0028b